+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Braintree iOS SDK Release Notes
 
+## unreleased
+* BraintreePayPal
+  * Add `BTContactInformation` request object
+  * Add `BTPayPalCheckoutRequest.contactInformation` optional property
+
 ## 6.26.0 (2025-01-21)
 * BraintreePayPal
   * Fix bug to ensure that `BTPayPalVaultRequest.userAuthenticationEmail` is not sent as an empty string